// src/shared/api/schema.ts
export interface components {
    schemas: {
// ...
        ExpiringContractsChartResponse: {
            /** Withindays */
            withinDays: number;
            /** Totalcount */
            totalCount: number;
            /** Items */
            items: components["schemas"]["ExpiringContractsChartItemResponse"][];
        };
// ...
        FlatsByStatusChartItemResponse: {
            /** Status */
            status: string;
            /** Count */
            count: number;
        };
// ...
        OccupancyChartItemResponse: {
            /** Name */
            name: string;
            /** Occupancypercent */
            occupancyPercent: number;
            /** Totalunits */
            totalUnits: number;
        };
// ...
        PaymentStatusChartItemResponse: {
            /** Status */
            status: string;
            /** Count */
            count: number;
        };
        /** PaymentsDynamicChartItemResponse */
        PaymentsDynamicChartItemResponse: {
            /** Periodlabel */
            periodLabel: string;
            /** Paid */
            paid: number;
            /** Unpaid */
            unpaid: number;
        };
// ...
        RentIncomeChartItemResponse: {
            /** Periodlabel */
            periodLabel: string;
            revenue: components["schemas"]["MoneyResponse"];
        };
        /** RequestsDynamicChartItemResponse */
        RequestsDynamicChartItemResponse: {
            /** Periodlabel */
            periodLabel: string;
            /** Created */
            created: number;
            /** Closed */
            closed: number;
            /** Inprogress */
            inProgress: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}